seperated temp file

+++ src/utils/compressImage.ts
@@ -7,9 +7,10 @@ const gmInstance = gm.subClass({ imageMagick: true });
 interface Options {
 
 }
-export default async function compressImage(filename: string, dirPath: string, opts?: Options) {
+export default async function compressImage(filename: string, dirPath?: string, opts?: Options): Promise<string> {
 
   return await new Promise(async (res, rej) => {
+    if (!dirPath) return rej(undefined);
     const currentExt = path.extname(dirPath);
     if (currentExt !== ".webp" && currentExt !== ".gif") { 
       const newDir = path.join(path.dirname(dirPath), path.basename(dirPath, currentExt) + ".webp")
@@ -23,7 +24,7 @@ export default async function compressImage(filename: string, dirPath: string, o
       .quality(90)
       .autoOrient()
       .write(dirPath, err => {
-        if (err) {
+        if (err && dirPath) {
           deleteFile(dirPath);
           return rej(err);
         }
